Provide a wrapper around core_user_create_users webservice.
The wrapper is local_ws_association_online_create_users, and it is
exactly the same (delegates to the original function) except that it
allows users with the same emails as existing users to be created.

This is inline with the feature provided by the Moodle batch importer,
which allows people with the same email to be imported if a checkbox
is selected.